requirements.txt
flask
flask-cors
pyjwt
werkzeug
